# CBG-657 - Fix race on benchmark tempdir teardown