The benchmark teardown was trying to remove a directory while the async compression code was still writing new files.

## Before

```
$ go test -run=- -bench='^BenchmarkLogRotation$' -v ./base
goos: darwin
goarch: amd64
pkg: github.com/couchbase/sync_gateway/base
BenchmarkLogRotation/rotate:false-compress:false-bytes:0-12         	 2052745	       594 ns/op
BenchmarkLogRotation/rotate:false-compress:false-bytes:1024000-12   	    3427	    576742 ns/op
BenchmarkLogRotation/rotate:false-compress:false-bytes:102400000-12 	      31	  58018333 ns/op
BenchmarkLogRotation/rotate:true-compress:false-bytes:0-12          	    2532	    456456 ns/op
BenchmarkLogRotation/rotate:true-compress:false-bytes:1024000-12    	     891	   1297064 ns/op
BenchmarkLogRotation/rotate:true-compress:false-bytes:102400000-12  	      32	  58446069 ns/op
BenchmarkLogRotation/rotate:true-compress:true-bytes:0-12           	       0	              NaN ns/op
BenchmarkLogRotation/rotate:true-compress:true-bytes:1024000-12     	       0	              NaN ns/op
BenchmarkLogRotation/rotate:true-compress:true-bytes:102400000-12   	      37	  64226538 ns/op
--- FAIL: BenchmarkLogRotation
    require.go:794:
        	Error Trace:	logging_test.go:134
        	            				logging_test.go:150
        	            				benchmark.go:190
        	            				benchmark.go:230
        	            				asm_amd64.s:1357
        	Error:      	Received unexpected error:
        	            	unlinkat /var/folders/0h/l73hpmwj0vsbpgw93jj39yk40000gn/T/benchmark-logrotate: directory not empty
        	Test:       	BenchmarkLogRotation
    logging_test.go:134:
        	Error Trace:	logging_test.go:134
        	            				logging_test.go:150
        	            				benchmark.go:190
        	            				benchmark.go:320
        	            				asm_amd64.s:1357
        	Error:      	Received unexpected error:
        	            	unlinkat /var/folders/0h/l73hpmwj0vsbpgw93jj39yk40000gn/T/benchmark-logrotate: directory not empty
        	Test:       	BenchmarkLogRotation
    logging_test.go:134:
        	Error Trace:	logging_test.go:134
        	            				logging_test.go:150
        	            				benchmark.go:190
        	            				benchmark.go:320
        	            				asm_amd64.s:1357
        	Error:      	Received unexpected error:
        	            	unlinkat /var/folders/0h/l73hpmwj0vsbpgw93jj39yk40000gn/T/benchmark-logrotate: directory not empty
        	Test:       	BenchmarkLogRotation
FAIL
exit status 1
FAIL	github.com/couchbase/sync_gateway/base	13.793s
FAIL
```

## After

```
$ go test -run=- -bench='^BenchmarkLogRotation$' -v ./base
goos: darwin
goarch: amd64
pkg: github.com/couchbase/sync_gateway/base
BenchmarkLogRotation/rotate:false-compress:false-bytes:0-12         	 2115913	       566 ns/op
BenchmarkLogRotation/rotate:false-compress:false-bytes:1024000-12   	    3967	    500865 ns/op
BenchmarkLogRotation/rotate:false-compress:false-bytes:102400000-12 	      33	  51593562 ns/op
BenchmarkLogRotation/rotate:true-compress:false-bytes:0-12          	    3055	    424369 ns/op
BenchmarkLogRotation/rotate:true-compress:false-bytes:1024000-12    	    1161	   1072950 ns/op
BenchmarkLogRotation/rotate:true-compress:false-bytes:102400000-12  	      38	  50118386 ns/op
BenchmarkLogRotation/rotate:true-compress:true-bytes:0-12           	    1928	   1018210 ns/op
BenchmarkLogRotation/rotate:true-compress:true-bytes:1024000-12     	    1135	   1251166 ns/op
BenchmarkLogRotation/rotate:true-compress:true-bytes:102400000-12   	      38	  52278519 ns/op
PASS
ok  	github.com/couchbase/sync_gateway/base	19.910s
```